requirements.txt
flask
requests
numpy
tensorflow
azure-ai-translation-text
azure-cognitiveservices-speech